refactor: convert components to function export, improve mobile style, and standardize file naming
- Converted all components from const to function export for consistency
- Improved responsive styling and layout for smartphone view
- Separated FAQ item logic into dedicated component (faqItem)
- Renamed files to use lowercase initial camelCase convention
