+++ frontend/src/App.jsx
@@ -28,14 +28,14 @@ import ResetPassword from "@pages/auth/resetPassword";
 import ChangePassword from "@pages/auth/changePassword";
 import TermsAndConditions from "@pages/termsAndConditions/termsAndConditions";
 import PrivacyPolicy from "@pages/privacyPolicy/privacyPolicy";
-import Faq from "@components/FAQ/FAQ";
+import Faq from "@components/faq/faq";
 import Contribute from "@pages/contribute/contribute";
 import GithubAuth from "@pages/auth/githubAuth";
 import HandleOAuth from "@pages/auth/handleOAuth";
 import NoPage from "@pages/noPage/noPage";
-import TableView from "@pages/TableView";
+import TableView from "@pages/tableView/tableView";
 
-function App() {
+export default function App() {
   return (
     <AuthProvider>
       <BrowserRouter>
@@ -87,5 +87,3 @@ function App() {
     </AuthProvider>
   );
 }
-
-export default App;
