+++ frontend/src/components/FAQ/FAQ.jsx
@@ -1,43 +1,9 @@
 import MetaData from '@/pages/noPage/metaData';
-import { useState } from 'react';
 import BackToTopButton from '../utils/backToTopButton';
 import { motion } from 'framer-motion';
+import FaqItem from './faqItem';
 
-const FaqItem = ({ question, answer }) => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  return (
-    <div className='border-b border-gray-700'>
-      <button
-        className='w-full py-5 px-4 flex justify-between items-center hover:bg-purple-900/30 cursor-pointer'
-        onClick={() => setIsOpen(!isOpen)}
-      >
-        <span className='text-lg font-medium text-white'>{question}</span>
-        <svg
-          className={`w-6 h-6 transform ${
-            isOpen ? 'rotate-180' : ''
-          } transition-transform text-white`}
-          fill='none'
-          stroke='currentColor'
-          viewBox='0 0 24 24'
-        >
-          <path
-            strokeLinecap='round'
-            strokeWidth='2'
-            d='M19 9l-7 7-7-7'
-          />
-        </svg>
-      </button>
-      {isOpen && (
-        <div className='px-4 pb-5'>
-          <p className='text-gray-300'>{answer}</p>
-        </div>
-      )}
-    </div>
-  );
-};
-
-const Faq = () => {
+export default function Faq() {
   const faqData = [
     {
       question: 'What is TerraQuake API?',
@@ -100,7 +66,7 @@ const Faq = () => {
               Frequently Asked Questions.
               <div className='h-0.5 w-1/3 md:w-1/4 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
               Find quick answers to common questions about TerraQuake API.
             </p>
           </motion.div>
@@ -123,5 +89,3 @@ const Faq = () => {
     </>
   );
 };
-
-export default Faq;
