+++ frontend/src/pages/auth/changePassword.jsx
@@ -148,7 +148,7 @@ export default function ChangePassword() {
                   Change password.
                   <div className='h-0.5 w-1/3 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
                 </h1>
-                <p className='text-xl text-left text-white/70 max-w-7xl'>
+                <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
                   Enter your email address below to receive a secure link to
                   reset your password. Make sure to use the same email
                   associated with your account.
