+++ frontend/src/pages/auth/forgotPassword.jsx
@@ -111,7 +111,7 @@ export default function forgotPassword() {
               Forgot Password?
               <div className='h-0.5 w-1/3 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
                Enter your email and we’ll send you a password reset link if it matches an existing account.
             </p>
           </motion.div>
