+++ frontend/src/pages/auth/resetPassword.jsx
@@ -11,7 +11,7 @@ import MetaData from '../noPage/metaData';
 import BackToTopButton from '@/components/utils/backToTopButton';
 import { motion } from 'framer-motion';
 
-export default function resetPassword() {
+export default function ResetPassword() {
   const [showPassword, setShowPassword] = useState(false);
   const togglePassword = () => setShowPassword((prev) => !prev);
 
@@ -111,7 +111,7 @@ export default function resetPassword() {
               Reset Password.
               <div className='h-0.5 w-1/3 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
               Enter your new password below to securely reset your account
               access.
             </p>
