+++ frontend/src/pages/auth/signUp.jsx
@@ -155,7 +155,7 @@ export default function SignUp() {
               Create account.
               <div className='h-0.5 w-1/3 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
               Create your TerraQuake account to start exploring real seismic
               events, customize your experience, and join our interactive
               training platform.
