+++ frontend/src/pages/contribute/contribute.jsx
@@ -1,5 +1,4 @@
 import { motion } from 'framer-motion';
-import { useState } from 'react';
 import { FaCode, FaBug, FaBook, FaDiscord } from 'react-icons/fa';
 import { IoIosGitPullRequest } from 'react-icons/io';
 import BackToTopButton from '@/components/utils/backToTopButton';
@@ -12,74 +11,7 @@ import {
   PULL_REQUESTS,
 } from '@/data/CONTRIBUTE';
 import hacktoberfest from '@images/hacktoberfest.svg';
-
-const ContributionCard = ({ icon, title, description, link, linkText }) => {
-  const [transformStyle, setTransformStyle] = useState(
-    'rotateX(0deg) rotateY(0deg) scale(1)'
-  );
-
-  const handleMouseMove = (e) => {
-    const { clientX, clientY, currentTarget } = e;
-    const { left, top, width, height } = currentTarget.getBoundingClientRect();
-    const x = (clientX - left) / width - 0.5;
-    const y = (clientY - top) / height - 0.5;
-    const rotateX = -y * 15;
-    const rotateY = x * 15;
-    setTransformStyle(
-      `rotateX(${rotateX}deg) rotateY(${rotateY}deg) scale(1.05)`
-    );
-  };
-
-  const handleMouseLeave = () => {
-    setTransformStyle('rotateX(0deg) rotateY(0deg) scale(1)');
-  };
-
-  const cardVariants = {
-    hidden: { opacity: 0, y: 50, scale: 0.95 },
-    visible: {
-      opacity: 1,
-      y: 0,
-      scale: 1,
-      transition: { duration: 0.5, ease: 'easeOut' },
-    },
-  };
-
-  return (
-    <motion.div
-      variants={cardVariants}
-      className='w-full'
-      style={{ perspective: '1000px' }}
-      onMouseMove={handleMouseMove}
-      onMouseLeave={handleMouseLeave}
-    >
-      <motion.div
-        className='bg-slate-800/40 backdrop-blur-md border border-white/10 rounded-xl p-6 shadow-2xl h-full flex flex-col'
-        style={{
-          transform: transformStyle,
-          transition: 'transform 0.2s ease-out',
-        }}
-      >
-        <div className='mb-4 text-3xl bg-gradient-to-r from-pink-500 to-purple-500 text-white bg-clip-text w-fit'>
-          {icon}
-        </div>
-        <h3 className='text-xl font-bold mb-3 text-gray-100'>{title}</h3>
-        <p className='text-gray-400 mb-4 flex-grow'>{description}</p>
-        <a
-          href={link}
-          target='_blank'
-          rel='noopener noreferrer'
-          className='inline-block text-indigo-400 hover:text-indigo-300 transition-colors duration-200 font-semibold group'
-        >
-          {linkText}
-          <span className='inline-block transition-transform duration-200 group-hover:translate-x-1'>
-            {' '}
-            &rarr;
-          </span>
-        </a>
-      </motion.div>
-    </motion.div>
-  );
-};
+import { ContributionCard } from './contributeCard';
 
 export default function Contribute() {
   const containerVariants = {
@@ -128,7 +60,7 @@ export default function Contribute() {
               Build the Future of Seismic Tech.
               <div className='h-0.5 w-1/3 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-xl text-center md:text-left text-white/70 max-w-7xl'>
               TerraQuake is built by a global community. Whether you're a
               developer, a seismologist, or a data enthusiast, your contribution
               is vital. Let's make seismic data more accessible together.
