+++ frontend/src/pages/privacyPolicy/privacyPolicy.jsx
@@ -42,7 +42,7 @@ export default function PrivacyPolicy() {
               Privacy Policy.
               <div className='h-0.5 w-1/4 md:w-1/5 mx-auto bg-gradient-to-r from-pink-500 via-purple-500 to-violet-500 my-2 rounded-full' />
             </h1>
-            <p className='text-xl text-left text-white/70 max-w-7xl'>
+            <p className='text-md text-center md:text-left text-white/70 max-w-7xl'>
               Last updated: 22-09-2025.
             </p>
           </motion.div>
